requirements.txt
flask==2.1.2
yara-python==4.3.0
boto3==1.26.105
botocore==1.29.105
urllib3==1.26.12